Moved fp8 stuff to fp8.py
